requirements.txt
flask
Flask-Cors
Werkzeug
google.generativeai
python-dotenv
pdfplumber
celery
redis
celery